done registration for new user

+++ studybud/base/views.py
@@ -1,5 +1,11 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
+from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
+from django.contrib.auth.models import User
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.forms import UserCreationForm
 from .models import Room, Topic
 from .forms import RoomForm
 
@@ -8,6 +14,57 @@ from .forms import RoomForm
 #     {'id': 2, 'name': 'Design with me'},
 #     {'id': 3, 'name': 'Frontend developers'},
 # ]
+
+def loginPage(request):
+    page = 'login'
+    
+    # if the user is already logged in, redirect to home
+    if request.user.is_authenticated:
+        return redirect('home')
+    
+    if request.method == 'POST':
+        username = request.POST.get('username').lower()
+        password = request.POST.get('password')
+
+        try:
+            user = User.objects.get(username=username)
+        except:
+            messages.error(request, 'User does not exist!')
+    
+        user = authenticate(request, username=username, password=password)
+        
+        if user is not None:
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'Username or password does not exist!')
+    
+    context = {'page': page}
+    return render(request, 'base/login_register.html', context)
+
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+
+def registerPage(request):
+    form = UserCreationForm()
+    
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST) # get data from form
+        
+        if form.is_valid(): # check if data is valid
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'AN error occured during registration!')
+        
+    return render(request, 'base/login_register.html', {'form': form})
 
 
 def home(request):
@@ -32,7 +89,11 @@ def room(request, pk):
     return render(request, 'base/room.html', context)
 
 
+@login_required(login_url='login')
 def createRoom(request):
+    """
+    User ought to be logged in to create a room
+    """
     form = RoomForm()
     
     if request.method == 'POST':
@@ -45,9 +106,15 @@ def createRoom(request):
     return render(request, 'base/room_form.html', context)
 
 
+@login_required(login_url='login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room) # fill in the form with the data of the room
+   
+    # if user try to edit the room that is not his own
+    if request.user != room.host:
+        return HttpResponse('You are not allowed to edit this room!')
+        
    
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room) # replace the data of the room with the data in the form (POST data)
@@ -59,8 +126,13 @@ def updateRoom(request, pk):
     return render(request, 'base/room_form.html', context)
 
 
+@login_required(login_url='login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+    
+    # if user try to edit the room that is not his own
+    if request.user != room.host:
+        return HttpResponse('You are not allowed to delete this room!')
     
     if request.method == 'POST':
         room.delete()
